Split out types into seperate files
