+++ PivotalTrackerDotNet/Domain/Stories.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace PivotalTrackerDotNet.Domain
+{
+    public class Stories:List<Story>
+    {
+        
+    }
+}
