+++ PivotalTrackerDotNet/MemberShipService.cs
@@ -3,11 +3,6 @@ using System.Linq;
 using PivotalTrackerDotNet.Domain;
 
 namespace PivotalTrackerDotNet {
-    public interface IMembershipService
-    {
-        List<Person> GetMembers(int projectId);
-    }
-
     public class MembershipService : AAuthenticatedService, IMembershipService {
         const string MemberShipEndpoint = "projects/{0}/memberships";
         public MembershipService(string token)
